Initial login API
BASE URL for Kitsu changed because oauth doesn't follow on from the
/edge endpoint. The api for oauth also does not give a response in
JSONAPI format so a standard json convertor factory is needed for
KitsuApi.login() calls. The oauth endpoint also requires a different
Content-Type so a check had to be included in JsonHeaderInterceptor to
prevent it being overrriden (test included).