Mvp basic setup
Switching to use a different implementation (rather than a Fragment in a Dialog, use a DialogFramgent) as the previous way does not work correctly (first opening of the bottom sheet is fine, but any further openings cause crashes. I imagine the previous implementation required something doing a bit more with lifecycle stuff but since it's complex enough for the Android devs to write DialogFragment for us I'm not going to bother with it.

Can't actually get this to work, however and I can't seem to figure out what I'm doing wrong ;-;